+++ tests/posts.test.ts
@@ -1,3 +1,4 @@
+/// <reference types="jest" />
 import request from 'supertest';
 import { MongoMemoryServer } from 'mongodb-memory-server';
 import mongoose from 'mongoose';
@@ -22,7 +23,7 @@ afterEach(async () => {
   // clear database between tests
   const collections = mongoose.connection.collections;
   for (const key in collections) {
-    await collections[key].deleteMany({});
+    if (collections[key]) await collections[key].deleteMany({});
   }
 });
